feat: finish client crud

+++ backend/src/modules/client/repository/client-repository.ts
@@ -1,20 +1,31 @@
 import prisma from '../../../prisma/db';
-import { ClientParams } from '../service/client-service';
+import {
+  CreateClientParams,
+  UpdateClientParams,
+} from '../service/client-service';
 
 export class ClientRepository {
-  async getById(id: number) {
-    return await prisma.client.findUnique({ where: { id } });
+  async getById(id: number, userId: number) {
+    return await prisma.client.findUnique({ where: { id, userId } });
   }
 
-  async create(data: ClientParams) {
+  async create(data: CreateClientParams) {
     const client = await prisma.client.create({ data });
     return client;
   }
-  async update(id: number, data: ClientParams) {
-    const client = await prisma.client.update({ where: { id }, data });
+  async update(data: UpdateClientParams) {
+    const { id, userId, ...clientData } = data;
+    const client = await prisma.client.update({
+      where: { id, userId },
+      data: clientData,
+    });
     return client;
   }
-  async delete(id: number) {
-    await prisma.client.delete({ where: { id } });
+  async delete(id: number, userId: number) {
+    await prisma.client.delete({ where: { id, userId } });
+  }
+  async getUserClients(userId: number) {
+    const clients = await prisma.client.findMany({ where: { userId } });
+    return clients;
   }
 }
